+++ relay-event-schema/src/protocol/ourlog/container.rs
@@ -2,24 +2,26 @@ use serde::{Deserialize, Serialize};
 
 use crate::protocol::utils;
 
-/// Metadata supplied with an item container containing spans.
+/// Metadata supplied with an item container containing logs.
 ///
-/// See also: <https://develop.sentry.dev/sdk/telemetry/spans/span-protocol/#version-and-ingest_settings-properties>.
+/// See also: <https://develop.sentry.dev/sdk/telemetry/logs/#version-and-ingest_settings-properties>.
 ///
 /// # Versions
 ///
 /// ## Behaviour
 ///
-/// Different versions influence how spans are normalized and processed:
+/// Different versions influence how logs are normalized and processed:
 ///
 /// | Version |      client ip     | user agent |
 /// |---------|--------------------|------------|
-/// | none    | {{auto}}           | never      |
+/// | none    | {{auto}}           | always     |
 /// | 2       | {{auto}}, settings | settings   |
 ///
-/// For example in version 2, the client ip is inferred if either the client address attribute is
-/// set to `auto` or the ingest settings specify `infer_ip` to `auto`.
-/// The user agent is only considered if ingest settings specify `infer_user_agent` as `auto`.
+/// For example, in version 2:
+///  - the client ip is inferred if either the client address attribute is
+///    set to `{{auto}}` or the ingest settings specify `infer_ip` as `auto`.
+///  - the user agent is only inferred if ingest settings specify `infer_user_agent` as `auto`,
+///    while for older SDKs which do not send any ingest settings, the user agent is always automatically inferred.
 ///
 /// ## Schema Changelog
 ///
@@ -30,10 +32,14 @@ use crate::protocol::utils;
 pub struct ContainerMetadata {
     /// The metadata version
     ///
-    /// The version influences how spans are processed during ingestion.
+    /// The version influences how logs are processed during ingestion.
+    ///
+    /// Currently supported versions:
+    /// - `none`: no processing changes
+    /// - `2`: adds support for `ingest_settings`
     #[serde(skip_serializing_if = "Option::is_none")]
     pub version: Option<u16>,
-    /// Settings controlling parts of the ingestion of individual spans.
+    /// Settings controlling parts of the ingestion of individual logs.
     ///
     /// Supported since version `2`.
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -87,7 +93,7 @@ impl InferIp {
 /// from the HTTP header provided user agent.
 #[derive(Debug, Copy, Clone, Deserialize, Serialize)]
 pub enum InferUserAgent {
-    /// The user agent should be inferred from the http request submitting the spans.
+    /// The user agent should be inferred from the http request submitting the logs.
     #[serde(rename = "auto")]
     Auto,
     /// The user agent should never be inferred.
@@ -110,7 +116,7 @@ mod tests {
     fn test_valid_round_trip() {
         let m: ContainerMetadata = serde_json::from_str(
             r#"{
-            "version": 123, 
+            "version": 123,
             "ingest_settings": {
                 "infer_ip": "auto",
                 "infer_user_agent": "never"