+++ relay-server/src/processing/transactions/mod.rs
@@ -121,7 +121,7 @@ impl Processor for TransactionProcessor {
             profiles,
         };
 
-        Some(Managed::with_meta_from(envelope, work))
+        Some(Managed::with_meta_from_managed_envelope(envelope, work))
     }
 
     async fn process(