+++ tests/integration/test_spans_standalone.py
@@ -28,7 +28,8 @@ def lcp_cls_inp_differences(mode):
     """
     if mode == "legacy":
         attributes = {
-            "browser.name": {"type": "string", "value": "Chrome"},
+            # The legacy pipeline extracts this attribute from `sentry_tags`.
+            "sentry.browser.name": {"type": "string", "value": "Chrome"},
             # Legacy behaviour, new field is `sentry.segment.name`
             # Maybe this shouldn't exist since parent and segment information is also removed
             "sentry.transaction": {"type": "string", "value": "/insights/projects/"},
@@ -37,7 +38,7 @@ def lcp_cls_inp_differences(mode):
     else:
         attributes = {
             # We additionally extract the browser version for EAP items
-            "sentry.browser.version": {"type": "string", "value": "141.0.0"},
+            "browser.version": {"type": "string", "value": "141.0.0"},
             # New for EAP items
             "sentry.observed_timestamp_nanos": {
                 "type": "string",
@@ -145,7 +146,7 @@ def test_lcp_span(
                 "type": "string",
                 "value": "https://s1.sentry-cdn.com/../sentry-loader.svg",
             },
-            "sentry.browser.name": {"type": "string", "value": "Chrome"},
+            "browser.name": {"type": "string", "value": "Chrome"},
             "sentry.description": {"type": "string", "value": "<unknown>"},
             "sentry.environment": {"type": "string", "value": "prod"},
             "sentry.exclusive_time": {"type": "double", "value": 0.0},
@@ -305,7 +306,7 @@ def test_cls_span(
                 "value": "div.app-1azrk9k.etjky0h0 > AppContainer > BodyContainer > BaseFooter",
             },
             "cls.source.3": {"type": "string", "value": "<unknown>"},
-            "sentry.browser.name": {"type": "string", "value": "Chrome"},
+            "browser.name": {"type": "string", "value": "Chrome"},
             "sentry.description": {
                 "type": "string",
                 "value": "AppContainer > NavContent > MobileTopbar > StyledButton",
@@ -450,7 +451,7 @@ def test_inp_span(
         "attributes": {
             "client.address": {"type": "string", "value": "127.0.0.1"},
             "inp": {"type": "double", "value": 104.0},
-            "sentry.browser.name": {"type": "string", "value": "Chrome"},
+            "browser.name": {"type": "string", "value": "Chrome"},
             "sentry.description": {
                 "type": "string",
                 "value": "<unknown>",