+++ tests/integration/test_spansv2.py
@@ -24,9 +24,7 @@ def envelope_with_spans(*payloads: dict, trace_info=None, metadata=None) -> Enve
     envelope.add_item(
         Item(
             type="span",
-            payload=PayloadRef(
-                json={"items": payloads, **(metadata if metadata is not None else {})}
-            ),
+            payload=PayloadRef(json={"items": payloads, **(metadata or {})}),
             content_type="application/vnd.sentry.items.span.v2+json",
             headers={"item_count": len(payloads)},
         )
@@ -128,8 +126,8 @@ def test_spansv2_basic(
             "valid_int": {"value": 9223372036854775807, "type": "integer"},
             "invalid_int": None,
             "invalid": None,
-            "sentry.browser.name": {"type": "string", "value": "Python Requests"},
-            "sentry.browser.version": {"type": "string", "value": "2.32"},
+            "browser.name": {"type": "string", "value": "Python Requests"},
+            "browser.version": {"type": "string", "value": "2.32"},
             "sentry.dsc.environment": {"type": "string", "value": "prod"},
             "sentry.dsc.public_key": {
                 "type": "string",
@@ -248,7 +246,7 @@ def test_spansv2_trimming_basic(
             # This is sufficient for all builtin attributes not
             # to be trimmed. The span fields that aren't trimmed
             # also still count for the size limit.
-            "trimming": {"span": {"maxSize": 453}},
+            "trimming": {"span": {"maxSize": 439}},
         }
     )
 
@@ -323,8 +321,8 @@ def test_spansv2_trimming_basic(
                 "value": "This is actually a pretty long string",
             },
             "custom.invalid.attribute": None,
-            "sentry.browser.name": {"type": "string", "value": "Python Requests"},
-            "sentry.browser.version": {"type": "string", "value": "2.32"},
+            "browser.name": {"type": "string", "value": "Python Requests"},
+            "browser.version": {"type": "string", "value": "2.32"},
             "sentry.dsc.environment": {"type": "string", "value": "prod"},
             "sentry.dsc.public_key": {
                 "type": "string",
@@ -344,7 +342,7 @@ def test_spansv2_trimming_basic(
         },
         "_meta": {
             "attributes": {
-                "": {"len": 505},
+                "": {"len": 491},
                 "custom.array.attribute": {
                     "value": {
                         "2": {