+++ tests/integration/test_vercel_logs.py
@@ -66,7 +66,6 @@ EXPECTED_ITEMS = [
         "timestamp": mock.ANY,
         "attributes": {
             "vercel.id": {"stringValue": "1573817187330377061717300000"},
-            "sentry.browser.version": {"stringValue": "2.32"},
             "sentry.origin": {"stringValue": "auto.log_drain.vercel"},
             "server.address": {"stringValue": "my-app-abc123.vercel.app"},
             "vercel.source": {"stringValue": "build"},
@@ -80,8 +79,7 @@ EXPECTED_ITEMS = [
                 "intValue": time_within_delta(expect_resolution="ns")
             },
             "vercel.build_id": {"stringValue": "bld_cotnkcr76"},
-            "sentry.payload_size_bytes": {"intValue": "436"},
-            "sentry.browser.name": {"stringValue": "Python Requests"},
+            "sentry.payload_size_bytes": {"intValue": mock.ANY},
             "vercel.project_id": {"stringValue": "gdufoJxB6b9b1fEqr1jUtFkyavUU"},
             "sentry._meta.fields.trace_id": {
                 "stringValue": '{"meta":{"":{"rem":[["trace_id.missing","s"]]}}}'
@@ -102,7 +100,6 @@ EXPECTED_ITEMS = [
         "timestamp": mock.ANY,
         "attributes": {
             "vercel.path": {"stringValue": "/api/users"},
-            "sentry.browser.version": {"stringValue": "2.32"},
             "vercel.proxy.scheme": {"stringValue": "https"},
             "vercel.entrypoint": {"stringValue": "api/index.js"},
             "vercel.proxy.user_agent": {"stringValue": '["Mozilla/5.0..."]'},
@@ -115,7 +112,6 @@ EXPECTED_ITEMS = [
             "vercel.execution_region": {"stringValue": "sfo1"},
             "sentry.severity_text": {"stringValue": "info"},
             "sentry.span_id": {"stringValue": "f24e8631bd11faa7"},
-            "sentry.browser.name": {"stringValue": "Python Requests"},
             "vercel.project_id": {"stringValue": "gdufoJxB6b9b1fEqr1jUtFkyavUU"},
             "vercel.request_id": {
                 "stringValue": "643af4e3-975a-4cc7-9e7a-1eda11539d90"
@@ -134,7 +130,7 @@ EXPECTED_ITEMS = [
             "sentry.timestamp_precise": {
                 "intValue": time_within_delta(expect_resolution="ns")
             },
-            "sentry.payload_size_bytes": {"intValue": "889"},
+            "sentry.payload_size_bytes": {"intValue": mock.ANY},
             "vercel.proxy.region": {"stringValue": "sfo1"},
         },
         "clientSampleRate": 1.0,
@@ -188,7 +184,7 @@ def test_vercel_logs_json_array(
             "org_id": 1,
             "outcome": 0,
             "project_id": 42,
-            "quantity": 1325,
+            "quantity": 1205,
         },
     ]
 
@@ -233,6 +229,6 @@ def test_vercel_logs_ndjson(
             "org_id": 1,
             "outcome": 0,
             "project_id": 42,
-            "quantity": 1325,
+            "quantity": 1205,
         },
     ]